\name{dental}
\docType{data}
\alias{dental}
\title{Individual Dental Claims Data Set}
\description{
  Basic dental claims on a policy with a deductible of 50.
}
\usage{dental}
\format{A vector containing 10 observations}
\source{Klugman, S. A., Panjer, H. H. and Willmot, G. E. (1998),
  \emph{Loss Models, From Data to Decisions}, Wiley.
}
\keyword{datasets}
