\name{gdental}
\docType{data}
\alias{gdental}
\title{Grouped Dental Claims Data Set}
\description{
  Grouped dental claims, that is presented in a number of claims per
  claim amount group form.
}
\usage{gdental}
\format{
  An object of class \code{"grouped.data"} (inheriting from class
  \code{"data.frame"}) consisting of 10 rows and 2 columns. The
  environment of the object contains the plain vector of \code{cj} of
  group boundaries
}
\source{Klugman, S. A., Panjer, H. H. and Willmot, G. E. (1998),
  \emph{Loss Models, From Data to Decisions}, Wiley.
}
\seealso{
  \code{\link{grouped.data}} for a description of grouped data objects.
}
\keyword{datasets}
