\name{hachemeister}
\docType{data}
\alias{hachemeister}
\title{Hachemeister data set}
\description{
  Hachemeister (1975) data set giving average claim amounts in
  automobile liability insurance in five U.S. states between July 1970
  and June 1973 and the corresponding number of claims.
}
\usage{data(hachemeister)}
\format{A list with two components:
  \describe{
    \item{\code{claims}}{a \eqn{5 \times 12} matrix of average claim
      amounts}
    \item{\code{weights}}{a \eqn{5 \times 12} matrix of corresponding
      weights}
  }
}
\source{Hachemeister, C. A. (1975), Credibility for regression
  models with application to trend, \emph{Proceedings of the Berkeley
    Actuarial Research Conference on Credibility}, Academic Press.
}
\keyword{datasets}
  