% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xround.R
\name{xround}
\alias{xround}
\title{Xround}
\usage{
xround(x, digits = 0, half_up = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{integer indicating the number of decimal places}

\item{half_up}{Boolean indicating rounding method to be used. Default is TRUE.}
}
\value{
numeric vector
}
\description{
Convenience function for rounding numbers
}
