\name{hachemeister}
\docType{data}
\alias{hachemeister}
\title{Hachemeister data set}
\description{
  Hachemeister (1975) data set giving average claim amounts in
  automobile liability insurance in five U.S. states between July 1970
  and June 1973 and the corresponding number of claims.
}
\usage{hachemeister}
\format{
  A matrix with 5 rows and the following 25 columns:
  \describe{
    \item{\code{state}}{the state number;}
    \item{\code{ratio.1}, \dots, \code{ratio.12}}{the average claim
      amounts;}
    \item{\code{weight.1}, \dots, \code{weight.12}}{the 
      corresponding number of claims.}
  }
}
\source{
  Hachemeister, C. A. (1975), \emph{Credibility for regression
    models with application to trend}, Proceedings of the Berkeley
  Actuarial Research Conference on Credibility, Academic Press.
}
\keyword{datasets}
